Move methods to new Util/Strings class
move string methods from xml to strings class.
